minor change to overall notification system. still trying to figure out scheduling
